third_party/blink/web_tests/fast/js/pic/cached-named-property-getter-expected.txt

CONSOLE ERROR: Uncaught TypeError: RegExp.prototype.ignoreCase getter called on non-RegExp object
This page tests cached access to getters and setters. If the test passes, you'll see a series of PASS messages below.

PASS: testGetter({__proto__: {count: 'FAIL'}, get length(){ return this.count; }, count: 7}) should be 7 and is.
PASS: testGetter(testFunction3) should be 3 and is.
PASS: testGetter(testFunction5) should be 5 and is.
PASS: testProtoGetter({__proto__: {count: 'FAIL', get length(){ return this.count; }}, count: 7}) should be 7 and is.
PASS: testProtoGetter({__proto__: testFunction3, count: 'FAIL'}) should be 3 and is.
PASS: testProtoGetter({__proto__: testFunction5, count: 'FAIL'}) should be 5 and is.
PASS: testProtoChainGetter({__proto__: {__proto__: {count: 'FAIL', get length(){ return this.count; }}}, count: 7}) should be 7 and is.
PASS: testProtoChainGetter({__proto__: {__proto__: testFunction3}, count: 'FAIL'}) should be 3 and is.
PASS: testProtoChainGetter({__proto__: {__proto__: testFunction5}, count: 'FAIL'}) should be 5 and is.
